requirements.txt
selenium~=4.21.0
pycryptodome~=3.20.0
passlib~=1.7.4
requests~=2.32.2
base58~=2.1.1
gnupg~=2.3.1
tablib~=3.6.1
openpyxl~=3.1.2
PyQt5~=5.15.10
pypdf2~=3.0.1
pillow~=10.3.0
PyMySQL~=1.1.1
psycopg2~=2.9.9
kafka~=1.3.5
pika~=1.3.2
redis~=5.0.4
kazoo~=2.10.0
python-jenkins~=1.8.2
pyvim~=3.0.3
pyvmomi~=8.0.2.0.1
hdfs~=2.7.3
pymongo~=4.7.2
pymssql~=2.3.0
paramiko~=3.4.0